Implemented Health System UI, if player has max 3 hearts they start with 3, upgrade system WIP, enemy detection for player is fixed, added max travel distance before destroying itself for projectile, fixed minor issues with detections

+++ Assets/_Scripts/Combat/Damage/DamageData.cs
@@ -4,15 +4,15 @@ namespace _Scripts.Combat.Damage
 {
     public class DamageData
     {
-        public float Amount { get; private set; }
+        public int Amount { get; private set; }
         public GameObject Source { get; private set; }
 
-        public DamageData(float amount, GameObject source)
+        public DamageData(int amount, GameObject source)
         {
             Amount = amount;
             Source = source;
         }
         
-        public void SetAmount(float amount) => Amount = amount;
+        public void SetAmount(int amount) => Amount = amount;
     }
 }
